src/test/java/com/woita/didemo/controllers/ConstructorInjectedControllerTest.java
package com.woita.didemo.controllers;

import com.woita.didemo.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

/**
 * @author mcbrydr on 01/08/19
 */
public class ConstructorInjectedControllerTest {

    private ConstructorInjectedController constructorInjectedController;

    @Before
    public void setUp() {
        this.constructorInjectedController = new ConstructorInjectedController(new GreetingServiceImpl());
    }

    @Test
    public void testGreeting() {
        assertEquals(GreetingServiceImpl.HELLO_GURUS, this.constructorInjectedController.sayHello());
    }
}
